SFA3: portrait sprites; better error if user doesn't extract RAR in full

+++ palmod/Game/GameLoad.cpp
@@ -327,7 +327,7 @@ CGameClass* CGameLoad::LoadFile(int nGameFlag, CHAR* szLoadFile)
 
         if (isSafeToRunGame)
         {
-            OutGame = CreateGame(nGameFlag, CurrFile.GetLength(), nGameRule);
+            OutGame = CreateGame(nGameFlag, (UINT32)CurrFile.GetLength(), nGameRule);
             OutGame->SetLoadDir(szLoadFile);
 
             if (OutGame->LoadFile(&CurrFile, 0))
